tests/ambient-light/AmbientLightSensor_onstatechange.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>AmbientLightSensor Test: onstatechange</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="https://www.w3.org/TR/ambient-light/">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>

var state = "idle";
var operation;
let sensor = new AmbientLightSensor();

test(function() {
  assert_equals(sensor.state, "idle");
}, "The default sensor.state is idle.");

sensor.onstatechange = function(event) {
  test(function() {
    assert_equals(sensor.state, state);
  }, "The sensor.state change to " + state + " after " + operation + ".");
};

sensor.onerror = function(event) {
  test(function() {
    assert_unreached(event.error.name + ":" + event.error.message);
  });
};

state = "activating";
operation = "sensor.start()";
sensor.start();

state = "idle";
operation = "sensor.stop()";
sensor.stop();

</script>
